+++ lib/Subscriber/Phpcr/QuerySubscriber.php
@@ -12,7 +12,6 @@
 namespace Sulu\Component\DocumentManager\Subscriber\Phpcr;
 
 use PHPCR\Query\QueryInterface;
-use PHPCR\Query\QueryManagerInterface;
 use PHPCR\SessionInterface;
 use Sulu\Component\DocumentManager\Collection\QueryResultCollection;
 use Sulu\Component\DocumentManager\Event\QueryCreateBuilderEvent;
@@ -20,6 +19,7 @@ use Sulu\Component\DocumentManager\Event\QueryCreateEvent;
 use Sulu\Component\DocumentManager\Event\QueryExecuteEvent;
 use Sulu\Component\DocumentManager\Events;
 use Sulu\Component\DocumentManager\Query\Query;
+use Sulu\Component\DocumentManager\Query\QueryBuilderConverter;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -39,13 +39,25 @@ class QuerySubscriber implements EventSubscriberInterface
     private $eventDispatcher;
 
     /**
-     * @param SessionInterface $session
-     * @param EventDispatcherInterface $eventDispatcher
+     * @var QueryBuilderConverter
      */
-    public function __construct(SessionInterface $session, EventDispatcherInterface $eventDispatcher)
-    {
+    private $builderConverter;
+
+    /**
+     * @var string
+     */
+    private $queryBuilderClass;
+
+    public function __construct(
+        SessionInterface $session,
+        EventDispatcherInterface $eventDispatcher,
+        QueryBuilderConverter $builderConverter,
+        $queryBuilderClass = 'Doctrine\ODM\PHPCR\Query\Builder\QueryBuilder'
+    ) {
         $this->session = $session;
         $this->eventDispatcher = $eventDispatcher;
+        $this->builderConverter = $builderConverter;
+        $this->queryBuilderClass = $queryBuilderClass;
     }
 
     /**
@@ -92,16 +104,15 @@ class QuerySubscriber implements EventSubscriberInterface
     }
 
     /**
-     * TODO: We should reuse the PHPCR-ODM query builder here, see:
-     *       https://github.com/doctrine/phpcr-odm/issues/627.
+     * Instantiate a new query builder and set it on the event.
      *
      * @param QueryCreateBuilderEvent $event
-     *
-     * @throws \Exception
      */
     public function handleCreateBuilder(QueryCreateBuilderEvent $event)
     {
-        throw new \Exception('Not implemented');
+        $builder = new $this->queryBuilderClass();
+        $builder->setConverter($this->builderConverter);
+        $event->setQueryBuilder($builder);
     }
 
     /**
@@ -113,17 +124,15 @@ class QuerySubscriber implements EventSubscriberInterface
     {
         $query = $event->getQuery();
         $locale = $query->getLocale();
+        $primarySelector = $query->getPrimarySelector();
         $phpcrResult = $query->getPhpcrQuery()->execute();
 
         $event->setResult(
-            new QueryResultCollection($phpcrResult, $this->eventDispatcher, $locale, $event->getOptions())
+            new QueryResultCollection($phpcrResult, $this->eventDispatcher, $locale, $event->getOptions(), $primarySelector)
         );
     }
 
-    /**
-     * @return QueryManagerInterface
-     */
-    private function getQueryManager()
+    protected function getQueryManager()
     {
         return $this->session->getWorkspace()->getQueryManager();
     }