+++ tests/Unit/MetadataTest.php
@@ -15,6 +15,9 @@ use Sulu\Component\DocumentManager\Metadata;
 
 class MetadataTest extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * @param Metadata
+     */
     private $metadata;
 
     public function setUp()
@@ -48,5 +51,29 @@ class MetadataTest extends \PHPUnit_Framework_TestCase
 
         $this->metadata->setClass('\stdClass');
         $this->assertNotSame($reflection, $this->metadata->getReflectionClass());
+    }
+
+    /**
+     * It should return a specific field mapping.
+     */
+    public function testGetFieldMapping()
+    {
+        $this->metadata->addFieldMapping('hello', ['property' => 'foo']);
+        $this->assertContains(
+            [
+                'property' => 'foo',
+            ],
+            $this->metadata->getFieldMapping('hello')
+        );
+    }
+
+    /**
+     * It should throw an exception if the given mapping does not exist.
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testGetFieldMappingNonExisting()
+    {
+        $this->metadata->getFieldMapping('hello');
     }
 }